+++ utils/applicationExceptions.js
